Create CustomColorCodedTable.
