+++ src/gui/add_items/ItemsListView.java
@@ -76,87 +76,8 @@ public class ItemsListView extends JPanel {
 
 		tableModel = new CustomTableModel();
 
-		table = new JTable(tableModel) {
+		table = new CustomColorCodedTable(tableModel);
 
-			@Override
-			public Component prepareRenderer(TableCellRenderer renderer, int row, int column) {
-				Component c = super.prepareRenderer(renderer, row, column);
-				if (!isRowSelected(row)) {
-					switch (colorCodingMode) {
-						case BY_FRESHNESS:
-
-							int freshnessCol = table.getColumnModel().getColumnIndex("Food Freshness");
-
-							Object freshnessValue = getValueAt(row, freshnessCol);
-							String freshness = "";
-
-							if (freshnessValue instanceof GenericTag) {
-								freshness = ((GenericTag<FoodFreshness>) freshnessValue).toString();
-							} else if (freshnessValue != null) {
-								freshness = freshnessValue.toString();
-							}
-
-							switch (freshness) {
-								case "Expired":
-									c.setBackground(new Color(252, 156, 156));
-									break;
-								case "Near_Expiry":
-									c.setBackground(new Color(236, 236, 127));
-									break;
-								case "Fresh":
-									c.setBackground(new Color(145, 252, 145));
-									break;
-								default:
-									c.setBackground(Color.WHITE);
-									break;
-							}
-							break;
-						case BY_FOOD_GROUP:
-							int foodGroupCol = getTable().getColumnModel().getColumnIndex("Food Group");
-							Object foodGroupValue = getValueAt(row, foodGroupCol);
-
-							String foodGroup = "";
-
-							if (foodGroupValue instanceof GenericTag) {
-								foodGroup = ((GenericTag<FoodGroup>) foodGroupValue).toString();
-							} else if (foodGroupValue != null) {
-								foodGroup = foodGroupValue.toString();
-							}
-
-							switch (foodGroup) {
-								case "Grain":
-									c.setBackground(new Color(255, 235, 156));
-									break;
-								case "Protein":
-									c.setBackground(new Color(252, 156, 156));
-									break;
-								case "Vegetable":
-									c.setBackground(new Color(193,219,155));
-									break;
-								case "Fruit":
-									c.setBackground(new Color(195, 177, 225));
-									break;
-								case "Dairy":
-									c.setBackground(new Color(207,219,231));
-									break;
-								default:
-									c.setBackground(Color.WHITE);
-									break;
-							}
-							break;
-
-						case OFF:
-							c.setBackground(Color.WHITE);
-							break;
-					}
-				}else {
-					if (!isRowSelected(row)) {
-						c.setBackground(Color.WHITE);
-					}
-				}
-				return c;
-			}
-		};
 
 		// Define table columns
 		tableModel.addColumn("Name");
@@ -369,7 +290,9 @@ public class ItemsListView extends JPanel {
 				colorCodingMode = ColorCodingMode.OFF;
 				break;
 		}
-		getTable().repaint();
+		if (table instanceof CustomColorCodedTable) {
+			((CustomColorCodedTable) table).setColorCodingMode(colorCodingMode);
+		}
 	}
 
 	public Container getC() {
